fix(replicator): destination start handling
In 367568e it was attempted to improve container start handling by
firing an additional container state event, with an "exited" status,
before the error channel.

However, this breaks a fundamental contract at the container
level. If the error channel fires then the container state channel is
not expected to be handled. Therefore, the correct fix for replicator
state is in the replicator itself.