fix(mediaserver): out-of-docker API URL
When outside of Docker, if `--server-url` is set to a non-public host,
like a private IP address, then the mediaserver API cannot be connected
to.

It's probably safe to assume that Octoplex is always running on the
Docker host in this case, in which case hitting localhost should be more
reliable.

Fixes #20